vpc work
Added testing that shows
* creating a vpc
* with r53 records
* multiple subnets for clients and servers

i also moved to coalesce from concat as it's easier to manage SG stuff that way. i think it was previously done with concat to support modding the SG back when it wasn't moddable, but now it is so you kinda want to be able to override it when you're looking to use multiple SGs. otherwise you end up with weird stuff like your client with one SG saying it can access internet and another that says it can't on two different subnets :scream:

adding support for private ip + zones