add support for private ip indexing in the hosts file
Previously when using hosts without a public ip the hosts file would not contain an index file. This change adds support for using private IP as the index value instead. It also contains a test for a private vpc environment with no public addresses and no hostnames to ensure that we can properly test for this change in future.

I also changed redpanda to broker in more places to maintain consistency.