adding support for private ip + zones (#7)
* vpc work

Added testing that shows
* creating a vpc
* with r53 records
* multiple subnets for clients and servers

i also moved to coalesce from concat as it's easier to manage SG stuff that way. i think it was previously done with concat to support modding the SG back when it wasn't moddable, but now it is so you kinda want to be able to override it when you're looking to use multiple SGs. otherwise you end up with weird stuff like your client with one SG saying it can access internet and another that says it can't on two different subnets :scream:

adding support for private ip + zones

added public + coalesce w/ rp subnet

I figured users might like to have everything on a non default subnet (for example in this case) so I decided to set up the subnetting such that you can do one of:

* not define a subnet and get one for free
* define one subnet in redpanda_subnet_id and everything uses it
* define two or three separate subnets and have the relevant stuff use them

also tossed in an issue fix

https://github.com/redpanda-data/terraform-aws-redpanda-cluster/issues/6

* adds readmes + description cleanup

* rework to support multi az

The intention is to support the following potential workflows:
* default vpc+subnet, single az
* provided vpc + 1 shared subnet for all node types, single az
* provided vpc + individual subnet (that may or may not be shared) for each node type, single az
* provided vpc+ 1 shared subnet, multi az
* provided vpc+ individual subnet (that may or may not be shared) for each node type, multi az

We only support a max of one subnet per az per node type because that's our current scope and honestly a pretty reasonable limitation for the vast majority of use cases!

Subnet merging and lookups within try blocks are used to support all of our workflows. Empty subnets will first attempt to resolve to broker (allowing for shared subnets) before falling back on an empty quote (allowing for default subnets).

In addition I've also done some limited renaming. Redpanda instances are now known as Broker instances. Counts for nodes and clients have been changed to broker_count and client_count to improve clarity, readability and searchability.

Doc updates will come in a future commit.

* add support for private ip indexing in the hosts file

Previously when using hosts without a public ip the hosts file would not contain an index file. This change adds support for using private IP as the index value instead. It also contains a test for a private vpc environment with no public addresses and no hostnames to ensure that we can properly test for this change in future.

I also changed redpanda to broker in more places to maintain consistency.

* firewall rule updates and tf vars removed

The TF inputs and outputs occur in docs autogenerated by terraform using the descriptions we provide so duplicating it in the readme is making us do twice the work for no additional value.

* enable instances to reach s3